compleate switch to tui for anime
