--- src/ui/mod.rs
@@ -1,1 +1,0 @@
-pub mod anime_ui;
